// api/models/DeckListItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DeckListItem = {
    title: string;
    description?: (string | null);
    category: string;
    id: string;
    last_studied?: (string | null);
    studied_today: number;
    current_progress: number;
    total_items: number;
};